eXPERIMENTS/cvv maker/requirements.txt
fastapi
uvicorn
python-docx
jinja2
pdfkit
pillow
imgkit